refactor: Move controller under controller package
